pyproject.toml
[project]
name = "vidstich"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "moviepy>=2.2.1",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow<11",
    "scikit-image>=0.25.2",
]
